@@ -3,17 +3,19 @@
   "version": "0.0.3",
   "private": true,
   "devDependencies": {
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
     "concurrently": "6.4.0",
     "react-scripts": "4.0.3"
   },
   "dependencies": {
     "autoprefixer-stylus": "1.0.0",
-    "firebase": "8.10.0",
+    "firebase": "^9.8.2",
     "prop-types": "15.7.2",
     "re-base": "4.0.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-router-dom": "5.3.0",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-router-dom": "^6.3.0",
     "react-transition-group": "4.4.2",
     "serve": "13.0.2",
     "stylus": "0.55.0"
